# core/core-api/src/main/resources/static/docs/openapi3.yaml
openapi: 3.0.1
info:
  title: soo-auth-service API
  description: 인증 서비스 API 가이드 (REST API Docs).
  version: 0.0.1-SNAPSHOT
servers:
- url: http://localhost
...
paths:
...
  /api/v1/auth/local/logout:
    post:
      tags:
      - Local Auth
      summary: 로그아웃
      description: "현재 디바이스 또는 전체 디바이스 로그아웃을 처리합니다.\n\n### 인증\n- `Authorization: Bearer\
        \ {accessToken}` 필요\n\n### 헤더\n- `Authorization`: Bearer 액세스 토큰\n- `X-Device-Id`:\
        \ 디바이스 식별자\n\n### 요청 바디 필드\n- `refreshToken`: 리프레시 토큰 (단건 로그아웃)\n- `logoutAll`:\
        \ 전체 로그아웃 여부\n\n### 응답 바디 필드\n- `result`: 요청 처리 결과 (SUCCESS/ERROR)\n- `meta.timestamp`:\
        \ 응답 생성 시각 (서버 로컬 시간)\n- `meta.requestId`: 요청 추적 ID (있을 때만 반환)\n- `meta.durationMs`:\
        \ 처리 시간(ms)\n- `meta.request.path`: 요청 경로\n- `meta.request.method`: HTTP 메\
        서드\n- `meta.request.query`: 쿼리 문자열 (없으면 null)\n- `meta.paging`: 페이징 메타 정보\
        \ (있을 때만 반환)\n- `meta.locale`: 응답 로케일 (있을 때만 반환)\n- `error`: 오류 정보 (에러 응답에\
        서만 포함)\n- `data`: 결과\n- `data.result`: OK\n\n### 공통 에러\n- **401** 인증 실패/만료\
        /위조\n- **403** 권한 없음(권한 필요한 API)\n- **400** validation 실패(필드 형식/필수 누락)\n\n\
        > 실제 요청/응답 예시는 Swagger UI의 **Examples** 섹션에 테스트 실행 결과로 자동 포함됩니다."
      operationId: auth-local-logout
      parameters:
      - name: Authorization
        in: header
        description: Bearer 액세스 토큰
        required: true
        schema:
          type: string
        example: Bearer token
      - name: X-Device-Id
        in: header
        description: 디바이스 식별자
        required: true
        schema:
          type: string
        example: device-001
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api-v1-services-serviceCode-auth-local-logout1868317452'
            examples:
              auth-local-logout:
                value: "{\n  \"refreshToken\" : \"refresh-token\",\n  \"logoutAll\"\
                  \ : false\n}"
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api-v1-auth-local-logout-2020253012'
              examples:
                auth-local-logout:
                  value: "{\n  \"result\" : \"SUCCESS\",\n  \"meta\" : {\n    \"timestamp\"\
                    \ : \"2026-04-06T10:27:51.725674715\",\n    \"request\" : {\n\
                    \      \"path\" : \"/api/v1/auth/local/logout\",\n      \"method\"\
                    \ : \"POST\"\n    }\n  },\n  \"data\" : {\n    \"result\" : \"\
                    OK\"\n  }\n}"
  /api/v1/auth/local/signup:
    post:
      tags:
      - Local Auth
      summary: 로컬 회원가입
      description: "이메일/비밀번호 기반 로컬 계정을 생성합니다.\n\n### 인증\n- 인증 불필요\n\n### 요청 바디 필드\n\
        - `email`: 이메일\n- `password`: 비밀번호\n- `name`: 이름\n- `nickname`: 닉네임\n- `gender`:\
        \ 성별 (MALE/FEMALE)\n- `phoneNumber`: 휴대폰 번호\n- `phoneVerificationToken`: 휴\
        대폰 인증 완료 토큰\n- `locale`: 로케일 (기본값: ko-KR)\n- `profileImageUrl`: 프로필 이미지 URL\n\
        - `thumbnailImageUrl`: 썸네일 이미지 URL\n- `birthyear`: 출생연도 (yyyy)\n- `birthday`:\
        \ 생일 (MM-DD)\n\n### 공통 에러\n- **401** 인증 실패/만료/위조\n- **403** 권한 없음(권한 필요한 API)\n\
        - **400** validation 실패(필드 형식/필수 누락)\n\n> 실제 요청/응답 예시는 Swagger UI의 **Examples**\
        \ 섹션에 테스트 실행 결과로 자동 포함됩니다."
      operationId: auth-local-signup
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api-v1-auth-local-signup-108875888'
            examples:
              auth-local-signup:
                value: "{\n  \"email\" : \"user@example.com\",\n  \"password\" : \"\
                  P@ssw0rd!\",\n  \"name\" : \"홍길동\",\n  \"nickname\" : \"gildong\"\
                  ,\n  \"gender\" : \"MALE\",\n  \"phoneNumber\" : \"+82 10-1234-5678\"\
                  ,\n  \"phoneVerificationToken\" : \"verified-phone-token\",\n  \"\
                  locale\" : \"ko-KR\",\n  \"profileImageUrl\" : \"http://dev.soo.io/profile.png\"\
                  ,\n  \"thumbnailImageUrl\" : \"http://dev.soo.io/thumbnail.png\"\
                  ,\n  \"birthyear\" : \"1990\",\n  \"birthday\" : \"01-31\"\n}"
      responses:
        "200":
          description: "200"
  /api/v1/auth/local/withdraw:
    post:
      tags:
      - Local Auth
      summary: 회원 탈퇴
      description: "사용자를 소프트 삭제하고 인증정보/토큰을 정리합니다.\n\n### 인증\n- `Authorization: Bearer\
        \ {accessToken}` 필요\n\n### 헤더\n- `Authorization`: Bearer 액세스 토큰\n\n### 요청\
        \ 바디 필드\n- `reason`: 탈퇴 사유\n\n### 응답 바디 필드\n- `result`: 요청 처리 결과 (SUCCESS/ERROR)\n\
        - `meta.timestamp`: 응답 생성 시각 (서버 로컬 시간)\n- `meta.requestId`: 요청 추적 ID (있을\
        \ 때만 반환)\n- `meta.durationMs`: 처리 시간(ms)\n- `meta.request.path`: 요청 경로\n-\
        \ `meta.request.method`: HTTP 메서드\n- `meta.request.query`: 쿼리 문자열 (없으면 null)\n\
        - `meta.paging`: 페이징 메타 정보 (있을 때만 반환)\n- `meta.locale`: 응답 로케일 (있을 때만 반환)\n\
        - `error`: 오류 정보 (에러 응답에서만 포함)\n- `data`: 결과\n- `data.result`: OK\n\n### 공\
        통 에러\n- **401** 인증 실패/만료/위조\n- **403** 권한 없음(권한 필요한 API)\n- **400** validation\
        \ 실패(필드 형식/필수 누락)\n\n> 실제 요청/응답 예시는 Swagger UI의 **Examples** 섹션에 테스트 실행 결과\
        로 자동 포함됩니다."
      operationId: auth-local-withdraw
      parameters:
      - name: Authorization
        in: header
        description: Bearer 액세스 토큰
        required: true
        schema:
          type: string
        example: Bearer token
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api-v1-services-serviceCode-auth-local-withdraw-1115063761'
            examples:
              auth-local-withdraw:
                value: "{\n  \"reason\" : \"privacy\"\n}"
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api-v1-auth-local-logout-2020253012'
              examples:
                auth-local-withdraw:
                  value: "{\n  \"result\" : \"SUCCESS\",\n  \"meta\" : {\n    \"timestamp\"\
                    \ : \"2026-04-06T10:27:51.75535614\",\n    \"request\" : {\n \
                    \     \"path\" : \"/api/v1/auth/local/withdraw\",\n      \"method\"\
                    \ : \"POST\"\n    }\n  },\n  \"data\" : {\n    \"result\" : \"\
                    OK\"\n  }\n}"
...
  /api/v1/auth/local/login/phone:
    post:
      tags:
      - Local Auth
      summary: 휴대폰 로그인
      description: "휴대폰 인증 토큰으로 로그인하고 액세스/리프레시 토큰을 발급합니다.\n\n### 인증\n- 인증 불필요\n\n\
        ### 헤더\n- `X-Device-Id`: 디바이스 식별자\n\n### 요청 바디 필드\n- `phoneNumber`: 휴대폰 번호\
        \n- `phoneVerificationToken`: 휴대폰 인증 완료 토큰\n\n### 응답 바디 필드\n- `result`: 요청\
        \ 처리 결과 (SUCCESS/ERROR)\n- `meta.timestamp`: 응답 생성 시각 (서버 로컬 시간)\n- `meta.requestId`:\
        \ 요청 추적 ID (있을 때만 반환)\n- `meta.durationMs`: 처리 시간(ms)\n- `meta.request.path`:\
        \ 요청 경로\n- `meta.request.method`: HTTP 메서드\n- `meta.request.query`: 쿼리 문자열\
        \ (없으면 null)\n- `meta.paging`: 페이징 메타 정보 (있을 때만 반환)\n- `meta.locale`: 응답 로\
        케일 (있을 때만 반환)\n- `error`: 오류 정보 (에러 응답에서만 포함)\n- `data`: 토큰 정보\n- `data.accessToken`:\
        \ 액세스 토큰\n- `data.accessExpiresInSec`: 액세스 토큰 만료(초)\n- `data.refreshToken`:\
        \ 리프레시 토큰\n- `data.refreshExpiresInSec`: 리프레시 토큰 만료(초)\n\n### 공통 에러\n- **401**\
        \ 인증 실패/만료/위조\n- **403** 권한 없음(권한 필요한 API)\n- **400** validation 실패(필드 형식\
        /필수 누락)\n\n> 실제 요청/응답 예시는 Swagger UI의 **Examples** 섹션에 테스트 실행 결과로 자동 포함됩니다\
        ."
      operationId: auth-local-login-phone
      parameters:
      - name: X-Device-Id
        in: header
        description: 디바이스 식별자
        required: true
        schema:
          type: string
        example: device-001
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api-v1-auth-phone-login1536767407'
            examples:
              auth-local-login-phone:
                value: "{\n  \"phoneNumber\" : \"+821012345678\",\n  \"phoneVerificationToken\"\
                  \ : \"verified-phone-token\"\n}"
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api-v1-auth-phone-login-1150988368'
              examples:
                auth-local-login-phone:
                  value: "{\n  \"result\" : \"SUCCESS\",\n  \"meta\" : {\n    \"timestamp\"\
                    \ : \"2026-04-06T10:27:51.704639697\",\n    \"request\" : {\n\
                    \      \"path\" : \"/api/v1/auth/local/login/phone\",\n      \"\
                    method\" : \"POST\"\n    }\n  },\n  \"data\" : {\n    \"accessToken\"\
                    \ : \"access-token\",\n    \"accessExpiresInSec\" : 3600,\n  \
                    \  \"refreshToken\" : \"refresh-token\",\n    \"refreshExpiresInSec\"\
                    \ : 1209600\n  }\n}"
...
  /api/v1/auth/local/signup/phone:
    post:
      tags:
      - Local Auth
      summary: 전화번호 간편 회원가입
      description: "전화번호 인증 토큰으로 최소 정보 계정을 생성합니다.\n\n### 인증\n- 인증 불필요\n\n### 요청 바디\
        \ 필드\n- `phoneNumber`: 휴대폰 번호\n- `phoneVerificationToken`: 휴대폰 인증 완료 토큰\n\n\
        ### 공통 에러\n- **401** 인증 실패/만료/위조\n- **403** 권한 없음(권한 필요한 API)\n- **400** validation\
        \ 실패(필드 형식/필수 누락)\n\n> 실제 요청/응답 예시는 Swagger UI의 **Examples** 섹션에 테스트 실행 결과\
        로 자동 포함됩니다."
      operationId: auth-local-signup-phone
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api-v1-auth-phone-login1536767407'
            examples:
              auth-local-signup-phone:
                value: "{\n  \"phoneNumber\" : \"+82 10-1234-5678\",\n  \"phoneVerificationToken\"\
                  \ : \"verified-phone-token\"\n}"
      responses:
        "200":
          description: "200"
  /api/v1/auth/local/token/refresh:
    post:
      tags:
      - Local Auth
      summary: 토큰 재발급
      description: "리프레시 토큰으로 새 액세스/리프레시 토큰을 발급합니다.\n\n### 인증\n- 인증 불필요(리프레시 토큰 필요\
        )\n\n### 헤더\n- `X-Device-Id`: 디바이스 식별자\n\n### 요청 바디 필드\n- `refreshToken`:\
        \ 리프레시 토큰\n\n### 응답 바디 필드\n- `result`: 요청 처리 결과 (SUCCESS/ERROR)\n- `meta.timestamp`:\
        \ 응답 생성 시각 (서버 로컬 시간)\n- `meta.requestId`: 요청 추적 ID (있을 때만 반환)\n- `meta.durationMs`:\
        \ 처리 시간(ms)\n- `meta.request.path`: 요청 경로\n- `meta.request.method`: HTTP 메\
        서드\n- `meta.request.query`: 쿼리 문자열 (없으면 null)\n- `meta.paging`: 페이징 메타 정보\
        \ (있을 때만 반환)\n- `meta.locale`: 응답 로케일 (있을 때만 반환)\n- `error`: 오류 정보 (에러 응답에\
        서만 포함)\n- `data`: 토큰 정보\n- `data.accessToken`: 액세스 토큰\n- `data.accessExpiresInSec`:\
        \ 액세스 토큰 만료(초)\n- `data.refreshToken`: 리프레시 토큰\n- `data.refreshExpiresInSec`:\
        \ 리프레시 토큰 만료(초)\n\n### 공통 에러\n- **401** 인증 실패/만료/위조\n- **403** 권한 없음(권한 필요\
        한 API)\n- **400** validation 실패(필드 형식/필수 누락)\n\n> 실제 요청/응답 예시는 Swagger UI의\
        \ **Examples** 섹션에 테스트 실행 결과로 자동 포함됩니다."
      operationId: auth-local-refresh
      parameters:
      - name: X-Device-Id
        in: header
        description: 디바이스 식별자
        required: true
        schema:
          type: string
        example: device-001
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api-v1-auth-local-token-refresh1056639717'
            examples:
              auth-local-refresh:
                value: "{\n  \"refreshToken\" : \"refresh-token\"\n}"
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api-v1-auth-phone-login-1150988368'
              examples:
                auth-local-refresh:
                  value: "{\n  \"result\" : \"SUCCESS\",\n  \"meta\" : {\n    \"timestamp\"\
                    \ : \"2026-04-06T10:27:51.792193967\",\n    \"request\" : {\n\
                    \      \"path\" : \"/api/v1/auth/local/token/refresh\",\n    \
                    \  \"method\" : \"POST\"\n    }\n  },\n  \"data\" : {\n    \"\
                    accessToken\" : \"access-token\",\n    \"accessExpiresInSec\"\
                    \ : 3600,\n    \"refreshToken\" : \"refresh-token\",\n    \"refreshExpiresInSec\"\
                    \ : 1209600\n  }\n}"
...
  /api/v1/services/{serviceCode}/auth/local/logout:
    post:
      tags:
      - Auth
      summary: 서비스별 로그아웃
      description: "serviceCode 기준 경로에서 로그아웃을 수행합니다(단건/전체 옵션).\n\n### 인증\n- `Authorization:\
        \ Bearer {accessToken}` 필요\n- `serviceCode` 유효성/활성 상태 검증\n\n### 헤더\n- `Authorization`:\
        \ Bearer 액세스 토큰\n- `X-Device-Id`: 디바이스 식별자\n\n### Path 파라미터\n- `serviceCode`:\
        \ 서비스 코드 (예: SHOP, CRM, DEFAULT)\n\n### 요청 바디 필드\n- `refreshToken`: 리프레시 토\
        큰 (단건 로그아웃)\n- `logoutAll`: 전체 로그아웃 여부\n\n### 응답 바디 필드\n- `result`: 요청 처리\
        \ 결과 (SUCCESS/ERROR)\n- `meta.timestamp`: 응답 생성 시각 (서버 로컬 시간)\n- `meta.requestId`:\
        \ 요청 추적 ID (있을 때만 반환)\n- `meta.durationMs`: 처리 시간(ms)\n- `meta.request.path`:\
        \ 요청 경로\n- `meta.request.method`: HTTP 메서드\n- `meta.request.query`: 쿼리 문자열\
        \ (없으면 null)\n- `meta.paging`: 페이징 메타 정보 (있을 때만 반환)\n- `meta.locale`: 응답 로\
        케일 (있을 때만 반환)\n- `error`: 오류 정보 (에러 응답에서만 포함)\n- `data`: 로그아웃 결과\n- `data.result`:\
        \ 처리 결과(OK)\n- `data.serviceCode`: 적용 서비스 코드\n\n### 공통 에러\n- **401** 인증 실패\
        /만료/위조\n- **403** 권한 없음(권한 필요한 API)\n- **400** validation 실패(필드 형식/필수 누락)\n\
        \n> 실제 요청/응답 예시는 Swagger UI의 **Examples** 섹션에 테스트 실행 결과로 자동 포함됩니다."
      operationId: auth-service-local-logout
      parameters:
      - name: serviceCode
        in: path
        description: "서비스 코드 (예: SHOP, CRM, DEFAULT)"
        required: true
        schema:
          type: string
      - name: Authorization
        in: header
        description: Bearer 액세스 토큰
        required: true
        schema:
          type: string
        example: Bearer token
      - name: X-Device-Id
        in: header
        description: 디바이스 식별자
        required: true
        schema:
          type: string
        example: device-001
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api-v1-services-serviceCode-auth-local-logout1868317452'
            examples:
              auth-service-local-logout:
                value: "{\n  \"refreshToken\" : \"refresh-token\",\n  \"logoutAll\"\
                  \ : false\n}"
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api-v1-services-serviceCode-auth-local-logout-879006433'
              examples:
                auth-service-local-logout:
                  value: "{\n  \"result\" : \"SUCCESS\",\n  \"meta\" : {\n    \"timestamp\"\
                    \ : \"2026-03-20T21:07:34.748601321\",\n    \"request\" : {\n\
                    \      \"path\" : \"/api/v1/services/SHOP/auth/local/logout\"\
                    ,\n      \"method\" : \"POST\"\n    }\n  },\n  \"data\" : {\n\
                    \    \"result\" : \"OK\",\n    \"serviceCode\" : \"SHOP\"\n  }\n\
                    }"
  /api/v1/services/{serviceCode}/auth/local/signup:
    post:
      tags:
      - Auth
      summary: 서비스별 로컬 회원가입
      description: "serviceCode 기준으로 로컬 계정을 생성합니다.\n\n### 인증\n- 인증 불필요\n- `serviceCode`\
        \ 유효성/활성 상태 검증\n\n### Path 파라미터\n- `serviceCode`: 서비스 코드 (예: SHOP, CRM, DEFAULT)\n\
        \n### 요청 바디 필드\n- `email`: 이메일\n- `password`: 비밀번호\n- `name`: 이름\n- `nickname`:\
        \ 닉네임\n- `gender`: 성별 (MALE/FEMALE)\n- `phoneNumber`: 휴대폰 번호\n- `locale`:\
        \ 로케일 (기본값: ko-KR)\n- `profileImageUrl`: 프로필 이미지 URL\n- `thumbnailImageUrl`:\
        \ 썸네일 이미지 URL\n- `birthyear`: 출생연도 (yyyy)\n- `birthday`: 생일 (MM-DD)\n\n###\
        \ 응답 바디 필드\n- `result`: 요청 처리 결과 (SUCCESS/ERROR)\n- `meta.timestamp`: 응답 생\
        성 시각 (서버 로컬 시간)\n- `meta.requestId`: 요청 추적 ID (있을 때만 반환)\n- `meta.durationMs`:\
        \ 처리 시간(ms)\n- `meta.request.path`: 요청 경로\n- `meta.request.method`: HTTP 메\
        서드\n- `meta.request.query`: 쿼리 문자열 (없으면 null)\n- `meta.paging`: 페이징 메타 정보\
        \ (있을 때만 반환)\n- `meta.locale`: 응답 로케일 (있을 때만 반환)\n- `error`: 오류 정보 (에러 응답에\
        서만 포함)\n- `data`: 회원가입 결과\n- `data.result`: 처리 결과(OK)\n- `data.serviceCode`:\
        \ 적용 서비스 코드\n\n### 공통 에러\n- **401** 인증 실패/만료/위조\n- **403** 권한 없음(권한 필요한 API)\n\
        - **400** validation 실패(필드 형식/필수 누락)\n\n> 실제 요청/응답 예시는 Swagger UI의 **Examples**\
        \ 섹션에 테스트 실행 결과로 자동 포함됩니다."
      operationId: auth-service-local-signup
      parameters:
      - name: serviceCode
        in: path
        description: "서비스 코드 (예: SHOP, CRM, DEFAULT)"
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api-v1-services-serviceCode-auth-local-signup-255962728'
            examples:
              auth-service-local-signup:
                value: "{\n  \"email\" : \"user@example.com\",\n  \"password\" : \"\
                  P@ssw0rd!\",\n  \"name\" : \"홍길동\",\n  \"nickname\" : \"gildong\"\
                  ,\n  \"gender\" : \"MALE\",\n  \"phoneNumber\" : \"+82 10-1234-5678\"\
                  ,\n  \"locale\" : \"ko-KR\",\n  \"profileImageUrl\" : \"http://dev.soo.io/profile.png\"\
                  ,\n  \"thumbnailImageUrl\" : \"http://dev.soo.io/thumbnail.png\"\
                  ,\n  \"birthyear\" : \"1990\",\n  \"birthday\" : \"01-31\"\n}"
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api-v1-services-serviceCode-auth-local-signup-107839563'
              examples:
                auth-service-local-signup:
                  value: "{\n  \"result\" : \"SUCCESS\",\n  \"meta\" : {\n    \"timestamp\"\
                    \ : \"2026-03-20T21:07:34.786358892\",\n    \"request\" : {\n\
                    \      \"path\" : \"/api/v1/services/SHOP/auth/local/signup\"\
                    ,\n      \"method\" : \"POST\"\n    }\n  },\n  \"data\" : {\n\
                    \    \"result\" : \"OK\",\n    \"serviceCode\" : \"SHOP\"\n  }\n\
                    }"
  /api/v1/services/{serviceCode}/auth/local/withdraw:
    post:
      tags:
      - Auth
      summary: 서비스별 회원 탈퇴(소프트 삭제)
      description: "serviceCode 기준 경로에서 회원 탈퇴(소프트 삭제)를 처리합니다.\n\n### 인증\n- `Authorization:\
        \ Bearer {accessToken}` 필요\n- `serviceCode` 유효성/활성 상태 검증\n\n### 헤더\n- `Authorization`:\
        \ Bearer 액세스 토큰\n\n### Path 파라미터\n- `serviceCode`: 서비스 코드 (예: SHOP, CRM, DEFAULT)\n\
        \n### 요청 바디 필드\n- `reason`: 탈퇴 사유\n\n### 응답 바디 필드\n- `result`: 요청 처리 결과 (SUCCESS/ERROR)\n\
        - `meta.timestamp`: 응답 생성 시각 (서버 로컬 시간)\n- `meta.requestId`: 요청 추적 ID (있을\
        \ 때만 반환)\n- `meta.durationMs`: 처리 시간(ms)\n- `meta.request.path`: 요청 경로\n-\
        \ `meta.request.method`: HTTP 메서드\n- `meta.request.query`: 쿼리 문자열 (없으면 null)\n\
        - `meta.paging`: 페이징 메타 정보 (있을 때만 반환)\n- `meta.locale`: 응답 로케일 (있을 때만 반환)\n\
        - `error`: 오류 정보 (에러 응답에서만 포함)\n- `data`: 탈퇴 처리 결과\n- `data.result`: 처리 결과\
        (OK)\n- `data.serviceCode`: 적용 서비스 코드\n\n### 공통 에러\n- **401** 인증 실패/만료/위조\n\
        - **403** 권한 없음(권한 필요한 API)\n- **400** validation 실패(필드 형식/필수 누락)\n\n> 실제\
        \ 요청/응답 예시는 Swagger UI의 **Examples** 섹션에 테스트 실행 결과로 자동 포함됩니다."
      operationId: auth-service-local-withdraw
      parameters:
      - name: serviceCode
        in: path
        description: "서비스 코드 (예: SHOP, CRM, DEFAULT)"
        required: true
        schema:
          type: string
      - name: Authorization
        in: header
        description: Bearer 액세스 토큰
        required: true
        schema:
          type: string
        example: Bearer token
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api-v1-services-serviceCode-auth-local-withdraw-1115063761'
            examples:
              auth-service-local-withdraw:
                value: "{\n  \"reason\" : \"privacy\"\n}"
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api-v1-services-serviceCode-auth-local-withdraw1895331346'
              examples:
                auth-service-local-withdraw:
                  value: "{\n  \"result\" : \"SUCCESS\",\n  \"meta\" : {\n    \"timestamp\"\
                    \ : \"2026-03-20T21:07:34.766701121\",\n    \"request\" : {\n\
                    \      \"path\" : \"/api/v1/services/SHOP/auth/local/withdraw\"\
                    ,\n      \"method\" : \"POST\"\n    }\n  },\n  \"data\" : {\n\
                    \    \"result\" : \"OK\",\n    \"serviceCode\" : \"SHOP\"\n  }\n\
                    }"
...
  /api/v1/services/{serviceCode}/auth/local/signup/phone:
    post:
      tags:
      - API
      summary: auth-service-local-signup-phone
      description: "auth-service-local-signup-phone\n\n### 인증\n- 인증 정책 확인 필요\n\n###\
        \ Path 파라미터\n- `serviceCode`: 서비스 코드 (예: SHOP, CRM, DEFAULT)\n\n### 요청 바디\
        \ 필드\n- `phoneNumber`: 휴대폰 번호\n\n### 응답 바디 필드\n- `result`: 요청 처리 결과 (SUCCESS/ERROR)\n\
        - `meta.timestamp`: 응답 생성 시각 (서버 로컬 시간)\n- `meta.requestId`: 요청 추적 ID (있을\
        \ 때만 반환)\n- `meta.durationMs`: 처리 시간(ms)\n- `meta.request.path`: 요청 경로\n-\
        \ `meta.request.method`: HTTP 메서드\n- `meta.request.query`: 쿼리 문자열 (없으면 null)\n\
        - `meta.paging`: 페이징 메타 정보 (있을 때만 반환)\n- `meta.locale`: 응답 로케일 (있을 때만 반환)\n\
        - `error`: 오류 정보 (에러 응답에서만 포함)\n- `data`: 회원가입 결과\n- `data.result`: 처리 결과\
        (OK)\n- `data.serviceCode`: 적용 서비스 코드\n\n### 공통 에러\n- **401** 인증 실패/만료/위조\n\
        - **403** 권한 없음(권한 필요한 API)\n- **400** validation 실패(필드 형식/필수 누락)\n\n> 실제\
        \ 요청/응답 예시는 Swagger UI의 **Examples** 섹션에 테스트 실행 결과로 자동 포함됩니다."
      operationId: auth-service-local-signup-phone
      parameters:
      - name: serviceCode
        in: path
        description: "서비스 코드 (예: SHOP, CRM, DEFAULT)"
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api-v1-services-serviceCode-auth-local-signup-phone-616279009'
            examples:
              auth-service-local-signup-phone:
                value: "{\n  \"phoneNumber\" : \"+82 10-1234-5678\"\n}"
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api-v1-services-serviceCode-auth-local-signup-107839563'
              examples:
                auth-service-local-signup-phone:
                  value: "{\n  \"result\" : \"SUCCESS\",\n  \"meta\" : {\n    \"timestamp\"\
                    \ : \"2026-03-20T21:07:34.83722562\",\n    \"request\" : {\n \
                    \     \"path\" : \"/api/v1/services/SHOP/auth/local/signup/phone\"\
                    ,\n      \"method\" : \"POST\"\n    }\n  },\n  \"data\" : {\n\
                    \    \"result\" : \"OK\",\n    \"serviceCode\" : \"SHOP\"\n  }\n\
                    }"
  /api/v1/services/{serviceCode}/auth/local/token/refresh:
    post:
      tags:
      - Auth
      summary: 서비스별 토큰 재발급(Refresh)
      description: "serviceCode 기준 경로에서 Refresh Token으로 Access Token을 재발급합니다.\n\n\
        ### 인증\n- 인증 불필요(일반적으로 refreshToken 자체가 인증 수단)\n- `serviceCode` 유효성/활성 상태\
        \ 검증\n\n### 헤더\n- `X-Device-Id`: 디바이스 식별자\n\n### Path 파라미터\n- `serviceCode`:\
        \ 서비스 코드 (예: SHOP, CRM, DEFAULT)\n\n### 요청 바디 필드\n- `refreshToken`: 리프레시 토\
        큰\n\n### 응답 바디 필드\n- `result`: 요청 처리 결과 (SUCCESS/ERROR)\n- `meta.timestamp`:\
        \ 응답 생성 시각 (서버 로컬 시간)\n- `meta.requestId`: 요청 추적 ID (있을 때만 반환)\n- `meta.durationMs`:\
        \ 처리 시간(ms)\n- `meta.request.path`: 요청 경로\n- `meta.request.method`: HTTP 메\
        서드\n- `meta.request.query`: 쿼리 문자열 (없으면 null)\n- `meta.paging`: 페이징 메타 정보\
        \ (있을 때만 반환)\n- `meta.locale`: 응답 로케일 (있을 때만 반환)\n- `error`: 오류 정보 (에러 응답에\
        서만 포함)\n- `data`: 토큰 재발급 결과\n- `data.serviceCode`: 적용 서비스 코드\n- `data.tokens`:\
        \ 재발급 토큰\n- `data.tokens.accessToken`: 액세스 토큰\n- `data.tokens.accessExpiresInSec`:\
        \ 액세스 토큰 만료(초)\n- `data.tokens.refreshToken`: 리프레시 토큰\n- `data.tokens.refreshExpiresInSec`:\
        \ 리프레시 토큰 만료(초)\n\n### 공통 에러\n- **401** 인증 실패/만료/위조\n- **403** 권한 없음(권한 필요\
        한 API)\n- **400** validation 실패(필드 형식/필수 누락)\n\n> 실제 요청/응답 예시는 Swagger UI의\
        \ **Examples** 섹션에 테스트 실행 결과로 자동 포함됩니다."
      operationId: auth-service-local-refresh
      parameters:
      - name: serviceCode
        in: path
        description: "서비스 코드 (예: SHOP, CRM, DEFAULT)"
        required: true
        schema:
          type: string
      - name: X-Device-Id
        in: header
        description: 디바이스 식별자
        required: true
        schema:
          type: string
        example: device-001
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api-v1-auth-local-token-refresh1056639717'
            examples:
              auth-service-local-refresh:
                value: "{\n  \"refreshToken\" : \"refresh-token\"\n}"
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api-v1-services-serviceCode-auth-local-token-refresh416663598'
              examples:
                auth-service-local-refresh:
                  value: "{\n  \"result\" : \"SUCCESS\",\n  \"meta\" : {\n    \"timestamp\"\
                    \ : \"2026-03-20T21:07:34.818539679\",\n    \"request\" : {\n\
                    \      \"path\" : \"/api/v1/services/SHOP/auth/local/token/refresh\"\
                    ,\n      \"method\" : \"POST\"\n    }\n  },\n  \"data\" : {\n\
                    \    \"serviceCode\" : \"SHOP\",\n    \"tokens\" : {\n      \"\
                    accessToken\" : \"access-token\",\n      \"accessExpiresInSec\"\
                    \ : 3600,\n      \"refreshToken\" : \"refresh-token\",\n     \
                    \ \"refreshExpiresInSec\" : 1209600\n    }\n  }\n}"
...
components:
  schemas:
...
    api-v1-services-serviceCode-auth-local-logout1868317452:
      type: object
      properties:
        logoutAll:
          type: boolean
          description: 전체 로그아웃 여부
          nullable: true
        refreshToken:
          type: string
          description: 리프레시 토큰 (단건 로그아웃)
          nullable: true
...
    api-v1-auth-local-signup-108875888:
      required:
      - email
      - gender
      - password
      - phoneNumber
      - phoneVerificationToken
      type: object
      properties:
        birthday:
          type: string
          description: 생일 (MM-DD)
          nullable: true
        phoneVerificationToken:
          type: string
          description: 휴대폰 인증 완료 토큰
        thumbnailImageUrl:
          type: string
          description: 썸네일 이미지 URL
          nullable: true
        password:
          type: string
          description: 비밀번호
        phoneNumber:
          type: string
          description: 휴대폰 번호
        birthyear:
          type: string
          description: 출생연도 (yyyy)
          nullable: true
        gender:
          type: string
          description: 성별 (MALE/FEMALE)
        nickname:
          type: string
          description: 닉네임
          nullable: true
        name:
          type: string
          description: 이름
          nullable: true
        locale:
          type: string
          description: "로케일 (기본값: ko-KR)"
          nullable: true
        profileImageUrl:
          type: string
          description: 프로필 이미지 URL
          nullable: true
        email:
          type: string
          description: 이메일
...
    api-v1-auth-local-token-refresh1056639717:
      required:
      - refreshToken
      type: object
      properties:
        refreshToken:
          type: string
          description: 리프레시 토큰
...
    api-v1-services-serviceCode-auth-local-withdraw-1115063761:
      type: object
      properties:
        reason:
          type: string
          description: 탈퇴 사유
          nullable: true
...
    api-v1-auth-phone-login1536767407:
      required:
      - phoneNumber
      - phoneVerificationToken
      type: object
      properties:
        phoneVerificationToken:
          type: string
          description: 휴대폰 인증 완료 토큰
        phoneNumber:
          type: string
          description: 휴대폰 번호
...
    api-v1-auth-local-logout-2020253012:
      required:
      - data
      - result
      type: object
      properties:
        result:
          type: string
          description: 요청 처리 결과 (SUCCESS/ERROR)
        data:
          required:
          - result
          type: object
          properties:
            result:
              type: string
              description: OK
          description: 결과
        meta:
          required:
          - timestamp
          type: object
          properties:
            request:
              required:
              - method
              - path
              type: object
              properties:
                path:
                  type: string
                  description: 요청 경로
                method:
                  type: string
                  description: HTTP 메서드
                query:
                  type: string
                  description: 쿼리 문자열 (없으면 null)
                  nullable: true
            requestId:
              type: string
              description: 요청 추적 ID (있을 때만 반환)
              nullable: true
            paging:
              type: object
              description: 페이징 메타 정보 (있을 때만 반환)
              nullable: true
            locale:
              type: string
              description: 응답 로케일 (있을 때만 반환)
              nullable: true
            durationMs:
              type: number
              description: 처리 시간(ms)
              nullable: true
            timestamp:
              type: string
              description: 응답 생성 시각 (서버 로컬 시간)
        error:
          type: object
          description: 오류 정보 (에러 응답에서만 포함)
          nullable: true
...